fix(map): address Burning Man Android review
