+++ androidApp/src/google/kotlin/org/meshtastic/app/map/MapView.kt
@@ -81,7 +81,6 @@ import com.google.maps.android.compose.ComposeMapColorScheme
 import com.google.maps.android.compose.GoogleMap
 import com.google.maps.android.compose.MapEffect
 import com.google.maps.android.compose.MapProperties
-import com.google.maps.android.compose.MapType
 import com.google.maps.android.compose.MapUiSettings
 import com.google.maps.android.compose.MapsComposeExperimentalApi
 import com.google.maps.android.compose.MarkerComposable
@@ -292,9 +291,7 @@ fun MapView(
         if (mode is GoogleMapMode.Main) mapViewModel.cameraPositionState else rememberCameraPositionState()
     val burningManTileProvider =
         remember(selectedBurningManPack?.file) {
-            selectedBurningManPack?.file?.let { file ->
-                runCatching { BurningManGoogleTileProvider(file) }.getOrNull()
-            }
+            selectedBurningManPack?.file?.let { file -> runCatching { BurningManGoogleTileProvider(file) }.getOrNull() }
         }
 
     if (mode is GoogleMapMode.Main) {
