+++ androidApp/src/google/kotlin/org/meshtastic/app/map/MapViewModel.kt
@@ -104,7 +104,6 @@ class MapViewModel(
 
     val selectedBurningManPack: StateFlow<SelectedBurningManPack?> = burningManPackRuntime.coordinator.selectedPack
 
-
     private val _selectedWaypointId = MutableStateFlow(savedStateHandle.get<Int>("waypointId"))
     val selectedWaypointId: StateFlow<Int?> = _selectedWaypointId.asStateFlow()
