+++ feature/map/src/androidHostTest/kotlin/org/meshtastic/feature/map/offline/BurningManPackRuntimeTest.kt
@@ -5,6 +5,14 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 package org.meshtastic.feature.map.offline
 
@@ -68,14 +76,13 @@ class BurningManPackRuntimeTest {
         }
     }
 
-    private fun installedRecord() =
-        BurningManPackRecord(
-            packId = "burning-man-2026",
-            sourceBuild = "20260720",
-            replicationTimestamp = "20260720",
-            installedAt = Instant.parse("2026-08-25T00:00:00Z"),
-            userSuppressed = false,
-        )
+    private fun installedRecord() = BurningManPackRecord(
+        packId = "burning-man-2026",
+        sourceBuild = "20260720",
+        replicationTimestamp = "20260720",
+        installedAt = Instant.parse("2026-08-25T00:00:00Z"),
+        userSuppressed = false,
+    )
 
     private fun writeValidatedPack(file: File) {
         file.parentFile?.mkdirs()
