+++ feature/map/src/androidMain/kotlin/org/meshtastic/feature/map/offline/BurningManPackCoordinator.kt
@@ -24,9 +24,9 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
+import java.io.File
 import kotlin.time.Clock
 import kotlin.time.Instant
-import java.io.File
 
 data class BurningManPackRecord(
     val packId: String,
@@ -52,29 +52,32 @@ class SharedPreferencesBurningManPackStore(context: Context) : BurningManPackSto
     private val preferences = context.getSharedPreferences(PREFERENCES_NAME, Context.MODE_PRIVATE)
 
     override fun load(): BurningManPackRecord? = preferences.getString(KEY_PACK_ID, null)?.let { packId ->
-            preferences.getString(KEY_INSTALLED_AT, null)?.let(Instant::parse)?.let { installedAt ->
-                BurningManPackRecord(
-                    packId = packId,
-                    sourceBuild = preferences.getString(KEY_SOURCE_BUILD, "") ?: "",
-                    replicationTimestamp = preferences.getString(KEY_REPLICATION_TIMESTAMP, "") ?: "",
-                    installedAt = installedAt,
-                    userSuppressed = preferences.getBoolean(KEY_USER_SUPPRESSED, false),
-                )
-            }
+        preferences.getString(KEY_INSTALLED_AT, null)?.let(Instant::parse)?.let { installedAt ->
+            BurningManPackRecord(
+                packId = packId,
+                sourceBuild = preferences.getString(KEY_SOURCE_BUILD, "") ?: "",
+                replicationTimestamp = preferences.getString(KEY_REPLICATION_TIMESTAMP, "") ?: "",
+                installedAt = installedAt,
+                userSuppressed = preferences.getBoolean(KEY_USER_SUPPRESSED, false),
+            )
         }
+    }
 
     override fun save(record: BurningManPackRecord?) {
-        preferences.edit().apply {
-            if (record == null) {
-                clear()
-            } else {
-                putString(KEY_PACK_ID, record.packId)
-                putString(KEY_SOURCE_BUILD, record.sourceBuild)
-                putString(KEY_REPLICATION_TIMESTAMP, record.replicationTimestamp)
-                putString(KEY_INSTALLED_AT, record.installedAt.toString())
-                putBoolean(KEY_USER_SUPPRESSED, record.userSuppressed)
+        preferences
+            .edit()
+            .apply {
+                if (record == null) {
+                    clear()
+                } else {
+                    putString(KEY_PACK_ID, record.packId)
+                    putString(KEY_SOURCE_BUILD, record.sourceBuild)
+                    putString(KEY_REPLICATION_TIMESTAMP, record.replicationTimestamp)
+                    putString(KEY_INSTALLED_AT, record.installedAt.toString())
+                    putBoolean(KEY_USER_SUPPRESSED, record.userSuppressed)
+                }
             }
-        }.apply()
+            .apply()
     }
 
     private companion object {
@@ -90,10 +93,9 @@ class SharedPreferencesBurningManPackStore(context: Context) : BurningManPackSto
 class BurningManPackCoordinator(
     private val filesDirectory: File,
     private val store: BurningManPackStore,
-    private val downloader: BurningManPackDownloader =
-        BurningManPackDownloader { bounds, destination ->
-            ProtomapsRegionDownloader(maxZoom = MAX_ZOOM).download(bounds, destination)
-        },
+    private val downloader: BurningManPackDownloader = BurningManPackDownloader { bounds, destination ->
+        ProtomapsRegionDownloader(maxZoom = MAX_ZOOM).download(bounds, destination)
+    },
 ) {
     private val stateMutex = Mutex()
     private val _selectedPack = MutableStateFlow<SelectedBurningManPack?>(null)
@@ -117,13 +119,14 @@ class BurningManPackCoordinator(
         stateMutex.withLock {
             destination.delete()
             val record =
-                store.load() ?: BurningManPackRecord(
-                    packId = PACK_ID,
-                    sourceBuild = "",
-                    replicationTimestamp = "",
-                    installedAt = Clock.System.now(),
-                    userSuppressed = true,
-                )
+                store.load()
+                    ?: BurningManPackRecord(
+                        packId = PACK_ID,
+                        sourceBuild = "",
+                        replicationTimestamp = "",
+                        installedAt = Clock.System.now(),
+                        userSuppressed = true,
+                    )
             store.save(record.copy(userSuppressed = true))
             _selectedPack.value = null
         }
@@ -132,10 +135,7 @@ class BurningManPackCoordinator(
     private suspend fun reconcileLocked(now: Instant, lastAuthorizedLocation: PackLocation?): SelectedBurningManPack? {
         var record = store.load()
         var selected = record?.let(::validatedSelection)
-        if (
-            record != null &&
-            (record.packId != PACK_ID || (selected == null && !record.userSuppressed))
-        ) {
+        if (record != null && (record.packId != PACK_ID || (selected == null && !record.userSuppressed))) {
             destination.delete()
             store.save(null)
             record = null
@@ -154,20 +154,22 @@ class BurningManPackCoordinator(
     }
 
     private suspend fun install(now: Instant): SelectedBurningManPack? = runCatching {
-            val downloadedPack = downloader.download(BOUNDS, destination)
-            val reader = PmtilesV3Reader(destination)
-            validatePack(reader)
-            val replicationTimestamp = requireNotNull(reader.metadata[REPLICATION_TIME_KEY]) {
+        val downloadedPack = downloader.download(BOUNDS, destination)
+        val reader = PmtilesV3Reader(destination)
+        validatePack(reader)
+        val replicationTimestamp =
+            requireNotNull(reader.metadata[REPLICATION_TIME_KEY]) {
                 "Burning Man pack has no replication timestamp"
             }
-            BurningManPackRecord(
-                packId = PACK_ID,
-                sourceBuild = downloadedPack.sourceBuild,
-                replicationTimestamp = replicationTimestamp,
-                installedAt = now,
-                userSuppressed = false,
-            )
-        }.fold(
+        BurningManPackRecord(
+            packId = PACK_ID,
+            sourceBuild = downloadedPack.sourceBuild,
+            replicationTimestamp = replicationTimestamp,
+            installedAt = now,
+            userSuppressed = false,
+        )
+    }
+        .fold(
             onSuccess = { record ->
                 val selected = SelectedBurningManPack(destination, record)
                 store.save(record)
@@ -189,20 +191,21 @@ class BurningManPackCoordinator(
     }
 
     private fun validatedSelection(record: BurningManPackRecord): SelectedBurningManPack? = runCatching {
-            require(record.packId == PACK_ID) { "Burning Man pack identifier does not match" }
-            require(destination.isFile) { "Burning Man pack is missing" }
-            val reader = PmtilesV3Reader(destination)
-            validatePack(reader)
-            require(reader.metadata[REPLICATION_TIME_KEY] == record.replicationTimestamp) {
-                "Burning Man pack replication timestamp does not match"
-            }
-            SelectedBurningManPack(destination, record)
-        }.getOrNull()
+        require(record.packId == PACK_ID) { "Burning Man pack identifier does not match" }
+        require(destination.isFile) { "Burning Man pack is missing" }
+        val reader = PmtilesV3Reader(destination)
+        validatePack(reader)
+        require(reader.metadata[REPLICATION_TIME_KEY] == record.replicationTimestamp) {
+            "Burning Man pack replication timestamp does not match"
+        }
+        SelectedBurningManPack(destination, record)
+    }
+        .getOrNull()
 
     private fun BurningManPackRecord.asManifest(): BurningManPackManifest = BurningManPackManifest(
-            packId = packId,
-            sourceBuild = sourceBuild,
-            installedAt = installedAt,
+        packId = packId,
+        sourceBuild = sourceBuild,
+        installedAt = installedAt,
         userSuppressed = userSuppressed,
     )
 
@@ -217,12 +220,6 @@ class BurningManPackCoordinator(
     private companion object {
         const val PACK_ID = "burning-man-2026"
         const val MAX_ZOOM = 15
-        val BOUNDS =
-            GeoBounds(
-                minLon = -119.287957,
-                minLat = 40.722536,
-                maxLon = -119.128520,
-                maxLat = 40.843420,
-            )
+        val BOUNDS = GeoBounds(minLon = -119.287957, minLat = 40.722536, maxLon = -119.128520, maxLat = 40.843420)
     }
 }
